test(api): move diffuseurs and search-document tests into mission suite
Follow the repo convention (one test file per controller): merge the two new
endpoint tests as describe blocks inside mission.test.ts instead of a separate
file named after the app-side usage.
